// src/broadcast/device/data/model/enum.ts
export enum Orientation {
    LANDSCAPE = 'LANDSCAPE',
    PORTRAIT = 'PORTRAIT',
  }
  
  export enum ChannelType {
    BASIC = 'BASIC',
    STANDARD = 'STANDARD',
    ADVANCED_HD = 'ADVANCED_HD',
    ADVANCED_SD = 'ADVANCED_SD',
  }

  export enum Resolution {
    R360 = 360,
    R480 = 480,
    R720 = 720,
    R1080 = 1080,
  }